document-api/pyproject.toml
[tool.poetry]
name = "document-api"
version = "0.1.0"
description = "FastAPI service for document upload and metadata management"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "*"}
python-multipart = "*"
httpx = "*"
pydantic = "*"
pydantic-settings = "*"
python-magic = "*"
opentelemetry-api = "^1.27.0"
opentelemetry-sdk = "^1.27.0"
opentelemetry-exporter-otlp = "^1.27.0"
opentelemetry-instrumentation-fastapi = "^0.48b0"
opentelemetry-instrumentation-httpx = "^0.48b0"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
